camera-stop-motion-bot/requirements.txt
gpiozero==1.6.2
picamera2==0.2.2
imageio==2.19.3
python-telegram-bot==13.12